2차 커밋

+++ KOSTA_3_final/src/main/java/com/kosta/KOSTA_3_final/model/user/Member.java
@@ -45,7 +45,7 @@ public class Member {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	Integer customer_id;
+	long customer_id;
 	@Column(name = "customer_name")
 	String customerName;
 	@Column(unique = true)
